+++ projects/gleks/ui/src/public-api.ts
@@ -4,6 +4,9 @@
 
 export * from './lib/components/button/button.component';
 export * from './lib/components/accordion/accordion.component';
+export * from './lib/components/collapsible/collapsible.component';
+export * from './lib/components/collapsible/collapsible-trigger.directive';
+export * from './lib/components/collapsible/collapsible-content.directive';
 export * from './lib/components/checkbox/checkbox.component';
 export * from './lib/components/dialog/confirmation-dialog/confirmation-dialog.component';
 export * from './lib/components/dialog/dialog.component';